% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{all_are_valid_colors}
\alias{all_are_valid_colors}
\title{Check if a string contains valid color strings}
\usage{
all_are_valid_colors(x)
}
\arguments{
\item{x}{A vector of colors in rgb or text string}
}
\description{
It appears I copied this function from https://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation
}
