% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{blend_palette}
\alias{blend_palette}
\title{Blend palette with a color}
\usage{
blend_palette(colors, blend, amount = 3)
}
\arguments{
\item{colors}{A vector of colors to blend}

\item{blend}{A color to blend with the colors}

\item{amount}{An integer between 0 and 10 to determine the strength of the blend}
}
\description{
This idea comes (and uses functions) from the \code{monochromeR} package and Cara
Thompson's talk titled \href{https://www.youtube.com/watch?v=tdaMw6NcNFM&t=2914s}{“Visualise, Optimise, Parameterise!”}.
The idea is that a set of colors in a palette will go together better if you
also blend a common color into all of them.
}
