% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_ab2gene}
\alias{card_ab2gene}
\title{Get gene information about an antibiotic}
\usage{
card_ab2gene(card, ab, return = "df")
}
\arguments{
\item{card}{output from \code{card_parse()}}

\item{ab}{character vector of antibiotic name(s)}

\item{return}{"df" to return a dataframe, "gene" to return a character vector of gene names}
}
\description{
Get gene information about an antibiotic
}
\seealso{
Other "CARD/RGI": 
\code{\link{card_ab_info}()},
\code{\link{card_gene2ab}()},
\code{\link{card_gene_info}()},
\code{\link{card_parse}()}
}
\concept{"CARD/RGI"}
