% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_ab_info}
\alias{card_ab_info}
\title{Get information about an antibiotic}
\usage{
card_ab_info(card, ab)
}
\arguments{
\item{card}{output from \code{card_parse()}}

\item{ab}{character vector of antibiotic name(s)}
}
\description{
Get information about an antibiotic
}
\seealso{
Other "CARD/RGI": 
\code{\link{card_ab2gene}()},
\code{\link{card_gene2ab}()},
\code{\link{card_gene_info}()},
\code{\link{card_parse}()}
}
\concept{"CARD/RGI"}
