% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_gene2ab}
\alias{card_gene2ab}
\title{Get gene information about an antibiotic}
\usage{
card_gene2ab(card, gene)
}
\arguments{
\item{card}{output from \code{card_parse()}}

\item{gene}{character vector of gene name(s)}
}
\description{
Get gene information about an antibiotic
}
\seealso{
Other "CARD/RGI": 
\code{\link{card_ab2gene}()},
\code{\link{card_ab_info}()},
\code{\link{card_gene_info}()},
\code{\link{card_parse}()}
}
\concept{"CARD/RGI"}
