% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_gene_info}
\alias{card_gene_info}
\title{Get antibiotic information about a gene}
\usage{
card_gene_info(card, gene)
}
\arguments{
\item{card}{output from \code{card_parse()}}

\item{gene}{character vector of gene name(s)}
}
\description{
Get antibiotic information about a gene
}
\seealso{
Other "CARD/RGI": 
\code{\link{card_ab2gene}()},
\code{\link{card_ab_info}()},
\code{\link{card_gene2ab}()},
\code{\link{card_parse}()}
}
\concept{"CARD/RGI"}
