% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_parse}
\alias{card_parse}
\title{Parse CARD database JSON file}
\usage{
card_parse(card_path, quiet = F)
}
\arguments{
\item{card_path}{Path to the card.json file}

\item{quiet}{If TRUE, suppress progress bar}
}
\value{
A dataframe with card annotations
}
\description{
This function parses the CARD JSON file and returns a dataframe. This
dataframe can be used by other functions in the function family.
}
\seealso{
Other "CARD/RGI": 
\code{\link{card_ab2gene}()},
\code{\link{card_ab_info}()},
\code{\link{card_gene2ab}()},
\code{\link{card_gene_info}()}
}
\concept{"CARD/RGI"}
