% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{cp_helper}
\alias{cp_helper}
\title{Internal helper function for parsing card.json file}
\usage{
cp_helper(card)
}
\arguments{
\item{card}{an inner card.json object}
}
\description{
This helper function is called in parallel by the \code{card_parse()} function,
and isn't intended to be accessed directly.
}
