% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EC_number_regex}
\alias{EC_number_regex}
\title{Regular expression for EC numbers}
\format{
A length-1 character vector.
}
\usage{
EC_number_regex
}
\description{
A regex pattern that matches Enzyme Commission (EC) numbers
of the form \code{1.2.3.4}.
}
\keyword{datasets}
