% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq.R
\name{fastq_filter_summary_amplicon}
\alias{fastq_filter_summary_amplicon}
\title{Summarize a fastq_filter.py file in amplicon mode}
\usage{
fastq_filter_summary_amplicon(file)
}
\arguments{
\item{file}{Output from fastq_filter.py}
}
\description{
Summarize a fastq_filter.py file in amplicon mode
}
