% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq.R
\name{fastq_filter_summary_meta}
\alias{fastq_filter_summary_meta}
\title{Summarize a fastq_filter.py file in metagenome mode}
\usage{
fastq_filter_summary_meta(file)
}
\arguments{
\item{file}{Output from fastq_filter.py}
}
\description{
Summarize a fastq_filter.py file in metagenome mode
}
