% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfm.R
\name{file2gfm}
\alias{file2gfm}
\title{Decorate a file string for gfm markdown in Obsidian}
\usage{
file2gfm(file, path = NA, verify = F)
}
\arguments{
\item{file}{A string representing a file name or path}

\item{path}{A string representing a file path}

\item{verify}{Logical indicating whether to verify the file exists}
}
\value{
A string with the file path as a gfm markdown link
}
\description{
This function takes a file path and returns a string with the file path as a
gfm formatted link that works in Obsidian. If a \code{path} is not given, then the
function will "expand" the \code{file} string to be a full path. This is meant to
expand "~/" to a full path, but can cause issues if you give just a file name
but no context as to the path.
}
