% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfm_link.R
\name{filepath2gfm}
\alias{filepath2gfm}
\title{Decorate a filepath string for gfm markdown in Obsidian}
\usage{
filepath2gfm(filepath, verify = F)
}
\arguments{
\item{filepath}{A string representing a file path}

\item{verify}{Logical indicating whether to verify the file exists}
}
\value{
A string with the file path as a gfm markdown link
}
\description{
Decorate a filepath string for gfm markdown in Obsidian
}
