% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{filter_asv_table}
\alias{filter_asv_table}
\title{Filter an ft object by minimum count and sample}
\arguments{
\item{x}{An ft object}

\item{min_count}{Minimum count for an ASV to be retained (default is 2)}

\item{min_sample}{Minimum number of samples an ASV must be present in to be retained (default is 2)}
}
\description{
Filter an ft object by minimum count and sample
}
