% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{ft}
\alias{ft}
\title{Feature table class}
\arguments{
\item{table}{A data frame containing the feature table with ASVs and sequences.}

\item{clusters}{A data frame containing cluster assignments, populated by \code{\link[=cluster_asv_table]{cluster_asv_table()}}.}

\item{filter}{A list containing filtering parameters applied to the table, populated by \code{\link[=filter_asv_table]{filter_asv_table()}}.}
}
\description{
This class represents a feature table (ft) object, which contains a table of ASVs and their sequences,
along with optional clustering information and filtering parameters.
}
