% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_symbol_regex}
\alias{gene_symbol_regex}
\title{Regular expression for gene symbols}
\format{
A length-1 character vector.
}
\usage{
gene_symbol_regex
}
\description{
A regex pattern that matches prokaryotic gene symbols
of the form \code{abcA1} (three lowercase letters, uppercase letter(s), optional digits).
}
\keyword{datasets}
