% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{get_alpha_diversity}
\alias{get_alpha_diversity}
\title{Alpha-Diversity Measures}
\usage{
get_alpha_diversity(p, id_col = "SAMPLE")
}
\arguments{
\item{p}{A phyloseq object}

\item{id_col}{The name of the column with sample IDs}
}
\description{
An extension of the phyloseq \code{estimate_richness()} function to include Shannon's
Evenness and sample depth. It also combines the normal estimate_richness output
with the sample data dataframe.
}
\details{
Although this new version removes much of the phyloseq code, running them instead
in Vegan or calculating manually, it does still require \code{phyloseq} from bioconductor,
which is not automatically installed with \code{jakR2} installation.
}
