% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{get_asv_table}
\alias{get_asv_table}
\title{Return ASV table from ft object}
\usage{
## S7 method for class <jakR2::ft>
get_asv_table(x)
}
\arguments{
\item{x}{An ft object}
}
\description{
Return ASV table from ft object
}
