% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_package_info}
\alias{get_package_info}
\title{Package info utility}
\usage{
get_package_info(attached = TRUE)
}
\arguments{
\item{attached}{Logical, whether to return only attached packages (default is TRUE)}
}
\description{
Package info utility
}
