% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{get_text_color}
\alias{get_text_color}
\title{Title}
\usage{
get_text_color(background_rgb, t = 0.5)
}
\arguments{
\item{background_rgb}{A hex color string (e.g. "#FFFFFF")}

\item{t}{A threshold value between 0 and 1 to determine the text color}
}
\description{
Title
}
