% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{get_tip_edges}
\alias{get_tip_edges}
\title{Get the edges that lead to a tip}
\usage{
get_tip_edges(tip_index, tree, type = "edges")
}
\arguments{
\item{tip_index}{}

\item{tree}{A phylo object}

\item{type}{what to return}
}
\description{
Get the edges that lead to a tip
}
