% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{get_tip_index}
\alias{get_tip_index}
\title{Get the index of a tip in a tree}
\usage{
get_tip_index(tip, tree)
}
\arguments{
\item{tip}{A character vector of tip labels}

\item{tree}{A phylo object}
}
\value{
A integer index of the tip label
}
\description{
Get the index of a tip in a tree
}
