% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{jak_grid}
\alias{jak_grid}
\title{Grid lines}
\usage{
jak_grid(grid_color = "transparent", border_width = 0.5)
}
\arguments{
\item{grid_color}{Color for major grid lines. Default \code{"transparent"} (no grid).}

\item{border_width}{Panel border line width. Default \code{0.5}.}
}
\value{
A ggplot2 theme object
}
\description{
Controls panel grid line appearance. Defaults to no grid lines.
}
