% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{jak_legend}
\alias{jak_legend}
\title{Legend styling}
\usage{
jak_legend(legend_position = "bottom", legend_size = 10)
}
\arguments{
\item{legend_position}{Legend position. One of \code{"bottom"}, \code{"top"}, \code{"left"},
\code{"right"}, or \code{"none"}. Default \code{"bottom"}.}

\item{legend_size}{Font size for legend text and title. Default \code{10}.}
}
\value{
A ggplot2 theme object
}
\description{
Sets legend position and text sizes.
}
