% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jak_palettes}
\alias{jak_palettes}
\title{JAK color palettes}
\format{
A named list of 23 character vectors, each containing hex color codes.
}
\usage{
jak_palettes
}
\description{
A named list of 23 color palettes used by \code{\link[=get_color_palette]{get_color_palette()}} and related functions.
}
\keyword{datasets}
