% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jak_samp}
\alias{jak_samp}
\title{Sample abundance matrix for examples and tests}
\format{
A matrix with 18 rows (features) and 20 columns (samples).
}
\usage{
jak_samp
}
\description{
A sample feature abundance matrix with 18 features and 20 samples.
}
\keyword{datasets}
