% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{jak_text}
\alias{jak_text}
\title{Text sizes and styles}
\usage{
jak_text(
  base_size = 10,
  plot_title_size = 12,
  axis_title_size = 11,
  plot_title_face = "plain",
  axis_title_face = "plain",
  base_family = ""
)
}
\arguments{
\item{base_size}{Base font size. Default \code{10}.}

\item{plot_title_size}{Plot title font size. Default \code{12}.}

\item{axis_title_size}{Axis title font size. Default \code{11}.}

\item{plot_title_face}{Plot title font face. Default "plain"}

\item{axis_title_face}{Axis title font face. Default "plain"}

\item{base_family}{Base font family. Default \code{""} (system default).}
}
\value{
A ggplot2 theme object
}
\description{
Sets font sizes, families, and text angles for plot elements.
}
