% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{jak_transparent}
\alias{jak_transparent}
\title{Transparent backgrounds}
\usage{
jak_transparent()
}
\value{
A ggplot2 theme object
}
\description{
Sets all background elements to transparent. Useful for plots intended
for slides or documents with varying background colors.
}
