% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jak_tree}
\alias{jak_tree}
\title{Sample phylogenetic tree for examples and tests}
\format{
A \code{phylo} object with 20 tips.
}
\usage{
jak_tree
}
\description{
A phylogenetic tree with 20 tips corresponding to the samples in \link{jak_samp}.
}
\keyword{datasets}
