% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfm.R
\name{obs2gfm}
\alias{obs2gfm}
\title{Decorate an Obsidian link for gfm markdown}
\usage{
obs2gfm(link, text = NA)
}
\arguments{
\item{link}{A string representing the link}

\item{text}{A string representing the text to display}
}
\description{
Decorate an Obsidian link for gfm markdown
}
