% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfm.R
\name{obs_callout}
\alias{obs_callout}
\title{Obsidian callout}
\usage{
obs_callout(type = "info", header = "", body = "")
}
\arguments{
\item{type}{The type of callout}

\item{header}{text to put in the header line}

\item{body}{text to put in the body}
}
\description{
Obisidan callouts can be one of many types. If a header line is not given,
then the header defaults to the name of the type. Text for an optional body
can also be provided.
}
