% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthofinder.R
\name{orthogroup_list2counts}
\alias{orthogroup_list2counts}
\title{Convert an orthogroup list dataframe to a count dataframe}
\usage{
orthogroup_list2counts(orthogroups, verify = F)
}
\arguments{
\item{orthogroups}{The orthogroup dataframe}

\item{verify}{If TRUE, return "counts" or "list" based on the input}
}
\description{
This function is used to convert the output of \code{read_orthogroups()} when the
type is set to "list" to a count format. Alternatively, it can run in verify
mode where it will return "counts" or "list" based on some basic inference.
}
