% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_jak}
\alias{palette_jak}
\title{A full-service color palette creator}
\usage{
palette_jak(p = "bay", colors = NULL, n = 2, reverse = FALSE)
}
\arguments{
\item{p}{A named color palette}

\item{colors}{An optional vector of rgb or colors to use. Overwrites \code{p}}

\item{n}{Number of colors to return}

\item{reverse}{A boolean for whether the palette should be reversed or not}
}
\description{
These palettes can be used all over.
}
