% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthofinder.R
\name{read_orthogroups}
\alias{read_orthogroups}
\title{Import an orthofinder TSV file}
\usage{
read_orthogroups(file, type = "counts", quiet = F)
}
\arguments{
\item{file}{The path to the Orthogroups/Orthogroups.tsv file in the standard orthofinder output}

\item{type}{Whether to return the counts or the lists of genes in each orthogroup}

\item{quiet}{If TRUE, suppress messages}
}
\description{
This code replaces the \code{jakR::orthofinder_summary()} function.
}
