% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_fill_jak_d}
\alias{scale_fill_jak_d}
\title{Custom discrete ggplot2 scale fill}
\usage{
scale_fill_jak_d(
  p = "bay",
  colors = NULL,
  order = "default",
  na_value = "gray50",
  name = ggplot2::waiver(),
  guide = "legend"
)
}
\arguments{
\item{p}{color palette to use from \code{jakR2::jak_palettes}}

\item{colors}{An optional vector of rgb or colors to use. Overwrites \code{p}}

\item{order}{Order of the palette, "default", "reverse" or "random"}

\item{na_value}{Color to use for NA values. Default \code{"gray50"}}

\item{name}{Legend title. Defaults to the name of the aesthetic (e.g., the variable name)}

\item{guide}{Guide to use for the legend. See \code{\link[ggplot2:guide_legend]{ggplot2::guide_legend()}} for details}
}
\description{
Custom discrete ggplot2 scale fill
}
