% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon.R
\name{seqtab_stats}
\alias{seqtab_stats}
\title{Extract DADA2 seqtab statistics}
\usage{
seqtab_stats(seqtab)
}
\arguments{
\item{seqtab}{A DADA2 sequence table}
}
\description{
Given a seqtab matrix, this function will calculate the counts and abundances
of ASVs and by their sequence lengths. \code{FREQ} is the number of ASVs, and \code{ABUNDANCE}
is the total number of reads for each sequence length.
}
\details{
A seqtab object itself is just a matrix, so the error checking can't always
determine if it is truly a seqtab object, or just another matrix. So,
}
