% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon.R
\name{seqtab_stats_bind}
\alias{seqtab_stats_bind}
\title{Stats of multiple seqtab objects}
\usage{
seqtab_stats_bind(...)
}
\arguments{
\item{...}{One or more seqtab objects}
}
\value{
A tibble with seqtab stats
}
\description{
Stats of multiple seqtab objects
}
