% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon.R
\name{seqtab_stats_plot}
\alias{seqtab_stats_plot}
\title{Plot multiple seqtab objects}
\usage{
seqtab_stats_plot(...)
}
\arguments{
\item{...}{One or more seqtab objects}
}
\value{
A ggplot object
}
\description{
Plot multiple seqtab objects
}
