% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{seqtab_to_ft}
\alias{seqtab_to_ft}
\title{Make ft object from DADA2 seqtab}
\usage{
seqtab_to_ft(seqtab, clean = "_F_filt.fastq.gz")
}
\arguments{
\item{seqtab}{A DADA2 sequence table (matrix) with ASVs as columns and samples as rows}

\item{clean}{A string to remove from the sample names in the feature table (default is "_F_filt.fastq.gz")}
}
\description{
Make ft object from DADA2 seqtab
}
