% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{show_color_palette}
\alias{show_color_palette}
\title{showColorPalette}
\usage{
show_color_palette(pal, labels = FALSE, label_angle = 0, alpha = 1)
}
\arguments{
\item{pal}{A color palette}

\item{labels}{Boolean for whether to show the rgb value as a label}

\item{label_angle}{The rotational angle for the label}

\item{alpha}{A numeric value between 0 and 1 for the transparency of the colors}
}
\description{
showColorPalette
}
