% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{all_are_valid_colors}
\alias{all_are_valid_colors}
\title{Check if a string contains valid color strings}
\usage{
all_are_valid_colors(x)
}
\arguments{
\item{x}{A vector of colors in rgb or text string}
}
\description{
This function is from \href{https://stackoverflow.com/questions/13289009/}{Stack Overflow #13289009}
originally by \href{https://stackoverflow.com/users/980833/josh-obrien}{Josh O'Brien}.
}
