% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{cluster_asv_table}
\alias{cluster_asv_table}
\title{Cluster ASV sequences from ft object}
\arguments{
\item{x}{An ft object}

\item{id}{Numeric value between 0 and 1 representing the clustering threshold (default is 0.99)}

\item{nproc}{Number of processors to use for parallel processing (default is 1)}

\item{quiet}{Logical, whether to suppress verbose output (default is TRUE)}
}
\description{
Cluster ASV sequences from ft object
}
