% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq.R
\name{fastq_info_summary}
\alias{fastq_info_summary}
\title{Summarize a fastq_info file}
\usage{
fastq_info_summary(file, fill = "cornflowerblue")
}
\arguments{
\item{file}{Path to output from fastq_info.py}

\item{fill}{Color to use for the histogram}
}
\description{
Summarize a fastq_info file
}
