% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{ft}
\alias{ft}
\title{Feature table class}
\usage{
ft(
  table = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names))
    {
         list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })(),
  clusters = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
         attr(out, "row.names") <- row.names
      
      out
     }
 })(),
  filter = list()
)
}
\arguments{
\item{table}{A data frame containing the feature table with ASVs and sequences.}
}
\description{
This class represents a feature table (ft) object, which contains a table of ASVs and their sequences,
along with optional clustering information and filtering parameters.
}
