% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{get_alpha_S}
\alias{get_alpha_S}
\title{Calculate richness (S) and sample depth}
\usage{
get_alpha_S(df, feature_id = "feature_id", id_col = "SAMPLE")
}
\arguments{
\item{df}{A dataframe with features as rows and samples as columns}

\item{feature_id}{The name of the column with feature IDs (default is "feature_id")}

\item{id_col}{The name of the column with sample IDs (default is "SAMPLE")}
}
\description{
Calculate richness (S) and sample depth
}
