% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{jak_theme}
\alias{jak_theme}
\title{JAK ggplot theme}
\usage{
jak_theme(
  base_size = 10,
  plot_title_size = 12,
  axis_title_size = 11,
  base_color = "gray30",
  plot_title_color = "gray30",
  axis_title_color = "gray30",
  bg_color = "transparent",
  grid_color = "gray80",
  border_width = 0.5,
  strip_size = 10,
  strip_color = "transparent",
  strip_line = 0,
  legend_size = 10,
  legend_position = "bottom",
  base_family = ""
)
}
\arguments{
\item{base_size}{Font size for most things}

\item{plot_title_size}{Font size for plot titles}

\item{axis_title_size}{Font size for axis titles}

\item{base_color}{Color for most text and lines}

\item{plot_title_color}{Color for plot titles}

\item{axis_title_color}{Color for axis titles}

\item{bg_color}{background color}

\item{grid_color}{Color for grid lines}

\item{strip_size}{Font size for facet strip titles}

\item{strip_color}{Color for facet strip titles}

\item{strip_line}{Line width for facet strip titles}

\item{legend_size}{Font size for legend}

\item{legend_position}{Position of legend}

\item{base_family}{Font family}
}
\description{
Modify theme of ggplots
}
