% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_jak}
\alias{palette_jak}
\title{A full-service color palette creator}
\usage{
palette_jak(p = "bay", colors = NULL, n = 2, order = "default")
}
\arguments{
\item{p}{A named color palette}

\item{colors}{An optional vector of rgb or colors to use. Overwrites \code{p}}

\item{n}{Number of colors to return}

\item{order}{default, reverse or random order}
}
\description{
The function creates a vector of RGB values suitable for ggplot or similar.
If using the \code{p} argument it will use one of the default palettes in the
\code{jak_palettes} data object, or you can pass it your own colors using the
\code{colors} argument. What makes this function useful is that it will extend
and interpolate the values of the given palette so it can return \code{n} colors.
The final parameter is whether you want to reverse the values or not.
}
\details{
Internally, the color palette is "ramped" using \code{colorRampPalette()}.
Further, colors passed via the \code{colors} argument are first checked whether
they're valid using the \code{all_are_valid_colors()} function.
}
