% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_fill_jak_d}
\alias{scale_fill_jak_d}
\title{Custom discrete ggplot2 scale fill}
\usage{
scale_fill_jak_d(p = "bay", colors = NULL, reverse = FALSE)
}
\arguments{
\item{p}{color palette to use from \code{jakR2::jak_palettes}}

\item{colors}{An optional vector of rgb or colors to use. Overwrites \code{p}}

\item{reverse}{A boolean for whether the palette should be reversed or not}
}
\description{
Custom discrete ggplot2 scale fill
}
